Move cursor construction into model
