+++ android-app/app/src/main/java/mozilla/voice/assistant/intents/communication/ui/contact/ContactViewModel.kt
@@ -1,10 +1,13 @@
 package mozilla.voice.assistant.intents.communication.ui.contact
 
 import android.app.Application
+import android.provider.ContactsContract
+import androidx.annotation.VisibleForTesting
 import androidx.lifecycle.AndroidViewModel
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
 import androidx.lifecycle.viewModelScope
+import androidx.loader.content.CursorLoader
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.withContext
@@ -53,5 +56,46 @@ class ContactViewModel(
 
     fun insert(contact: ContactEntity) = viewModelScope.launch(Dispatchers.IO) {
         repository.insert(contact)
+    }
+
+    @VisibleForTesting
+    internal fun comparisonStringsFor(name: String) =
+        arrayOf(
+            name, // just the nickname
+            "$name %", // first name
+            "% $name", // last name
+            "% $name %" // middle name
+        )
+
+    fun toCursorLoader(app: Application) =
+        CursorLoader(
+            app,
+            ContactsContract.Contacts.CONTENT_URI,
+            PROJECTION,
+            SELECTION,
+            comparisonStringsFor(nickname),
+            null
+        )
+
+    companion object {
+        private val PROJECTION: Array<out String> = arrayOf(
+            ContactsContract.Contacts._ID,
+            ContactsContract.Contacts.DISPLAY_NAME_PRIMARY,
+            ContactsContract.Contacts.PHOTO_THUMBNAIL_URI
+        )
+
+        // The following statements set up constants for a query to find contacts whose display
+        // name either match or contain a nickname provided at run-time.
+        private const val HAS_PHONE_TERM = "${ContactsContract.Contacts.HAS_PHONE_NUMBER} = 1"
+        private const val LIKE_TERM = "${ContactsContract.Contacts.DISPLAY_NAME_PRIMARY} LIKE ?"
+        private const val NUM_LIKE_TERMS = 4 // exact match, first name, last name, middle name
+        internal val SELECTION: String =
+            generateSequence { LIKE_TERM }
+                .take(NUM_LIKE_TERMS)
+                .joinToString(
+                    separator = " OR ",
+                    prefix = "$HAS_PHONE_TERM AND (",
+                    postfix = ")"
+                )
     }
 }
